// types.ts
export enum NotificationType {
    SUCCESS = 'success',
    ERROR = 'error',
    INFO = 'info',
}

export interface NotificationMessage {
    id: number;
    message: string;
    type: NotificationType;
}

export interface VersionInfo {
    version: string;
    releaseDate: string;
    changelog: string[];
}

export interface InstructionContent {
    version: string;
    title: string;
    content: string;
    lastUpdatedAt: Date;
}

export type ContactType = 'Phone' | 'WhatsApp' | 'Telegram' | 'Email';

export interface ContactInfo {
    id: string;
    type: ContactType;
    value: string;
    isPublic: boolean;
}

export interface SystemSettings {
    isCoursePopupEnabled: boolean;
    isInstructionPopupEnabled: boolean;
    logoUrl: string;
    contactInfo: ContactInfo[];
}

export interface Backup {
    id: string;
    timestamp: Date;
    data: string; // JSON string of the entire database state
}